+++ openapi/src/codegen.rs
@@ -698,6 +698,7 @@ pub fn gen_unions(out: &mut String, unions: &BTreeMap<String, InferredUnion>, me
 pub fn gen_variant_name(wire_name: &str, meta: &Metadata) -> String {
     match wire_name {
         "*" => "All".to_string(),
+        "self" => "Self_".to_string(),
         n => {
             if n.chars().next().unwrap().is_digit(10) {
                 format!("V{}", n.to_string().replace('-', "_").replace('.', "_"))
@@ -1039,7 +1040,7 @@ fn gen_field_type(
 ) -> String {
     match &field.schema_kind {
         // N.B. return immediately; if we want to use `Default` for bool rather than `Option`
-        SchemaKind::Type(Type::Boolean {}) => "bool".into(),
+        SchemaKind::Type(Type::Boolean(_)) => "bool".into(),
         SchemaKind::Type(Type::Number(_)) => "f64".into(),
         SchemaKind::Type(Type::Integer(format)) => {
             infer_integer_type(state, field_name, &format.format)
@@ -1290,6 +1291,11 @@ pub fn gen_field_rust_type<T: Borrow<Schema>>(
     // currency_options field is represented by an optional HashMap<String, T>, where the String is the currency code in ISO 4217 format.
     if field_name == "currency_options" {
         state.use_params.insert("CurrencyMap");
+
+        if ty.contains("CurrencyMap<") {
+            return ty;
+        }
+
         return format!("Option<CurrencyMap<{}>>", ty);
     }