+++ openapi/src/metadata.rs
@@ -4,7 +4,6 @@ use std::path::Path;
 
 use heck::{CamelCase, SnakeCase};
 use openapiv3::{ReferenceOr, SchemaKind};
-use tracing::trace;
 
 use crate::spec::{as_object_properties, Spec};
 use crate::{
@@ -210,6 +209,9 @@ pub fn metadata_requests<'a>(
             // special case for usage_records
             (_, _, Some("usage_records")) => "usage_records".to_string(),
 
+            // special case: terminal is a resource, not an object
+            (Some("terminal"), Some(x), _) => format!("terminal.{}", x),
+
             (Some(x), _, _) => x.to_string(),
             _ => {
                 // this should never happen
@@ -312,5 +314,5 @@ pub fn feature_groups() -> BTreeMap<&'static str, &'static str> {
 	]
 	.iter()
 	.copied()
-	.collect() 
+	.collect()
 }