+++ src/ids.rs
@@ -504,7 +504,7 @@ def_id!(CreditNoteLineItemId, "cnli_");
 def_id!(CustomerBalanceTransactionId, "cbtxn_");
 def_id!(CustomerId, "cus_");
 def_id!(DiscountId, "di_");
-def_id!(DisputeId, "dp_" | "du_");
+def_id!(DisputeId, "dp_" | "du_" | "pdp_");
 def_id!(EphemeralKeyId, "ephkey_");
 def_id!(EventId, "evt_");
 def_id!(FileId, "file_");
@@ -555,7 +555,7 @@ def_id!(
 def_id!(PersonId, "person_");
 def_id!(PlanId: String); // N.B. A plan id can be user-provided so can be any arbitrary string
 def_id!(PlatformTaxFeeId, "ptf");
-def_id!(PriceId, "price_");
+def_id!(PriceId, "price_" | "plan_"); // see #470
 def_id!(ProductId: String); // N.B. A product id can be user-provided so can be any arbitrary string
 def_id!(PromotionCodeId, "promo_");
 def_id!(QuoteId, "qt_");