+++ src/resources.rs
@@ -121,6 +121,7 @@ pub use {
         balance_transaction::*,
         billing_details::*,
         charge::*,
+        connect_account_reference::*,
         customer::*,
         custom_unit_amount::*,
         cash_balance::*,
@@ -136,6 +137,7 @@ pub use {
         payment_method_details_card_wallet_apple_pay::*,
         payment_method_details_card_wallet_google_pay::*,
         payment_method_options_customer_balance_eu_bank_account::*,
+        payment_method_options_us_bank_account_mandate_options::*,
         payment_method_config_biz_payment_method_configuration_details::*,
         payout::*,
         platform_tax_fee::*,
@@ -229,6 +231,7 @@ pub use {
         // need to import this afterwards so that the SubscriptionItemPriceDataRecurring
         // isn't silently ignored
         subscription::*,
+        subscriptions_trials_resource_trial_settings::*,
         subscription::PlanInterval as SubscriptionInterval,
         subscription::SubscriptionItemPriceDataRecurring as SubscriptionPriceDataRecurring,
         subscription::SubscriptionItemPriceData as SubscriptionPriceData,